Implements the SC Muse Client Code Generator
This is the initial implementation of the Sonos Control Muse Client
Code Generator. This uses the openapi-generator constructs and the
generate Muse OAS spec.

TODO:

- Today, the C++ generator does not support the `anyOf` construct from the
OAS spec. This causes problems for commands that we respond with an error
or a model T (for example the `createArea` command `event="area playerSetError"`).
For this implementation, I manually removed the use of `anyOf` from the OAS
spec.