Add oneOf property
Change oneOf property to a list of maps

I need a way to reference the data within the list elements from
mustache. I think that inserting a map is the easiest way. I could
create a new type, but I think that this works just as well.

For now, the only expected map key is "innerDataType."